Tried and failed to implement Ethash23 very generically
I wanted to implement Ethash23 very generically in order to play with different
storage types for the unsigned 32-bit ints we require without code duplication.

The effort failed because, for decent syntax, we need to wrap our type class
in ops, but the ops classes cannot be value classes (because they must be embedded
in the generic implementation to capture the types and the type class implementation,
but value classes can't be inner) and because it's laborious to write implicit classes
for type class implementors (effectively for context-bound arguments) because that
implies an illegal second constructor argument unless you extend the generic implementations
for each instance (to fill in the extra value in the superconstructor). Very annoying.