Downgrade to bouncycastle-1.51 so tests pass, will need to migrate from SHA3 to KECCAK
Updating to bouncycastle 1.53 breaks ethereum compatability for now, because the padding of the Keccak algorithm
was updated when it formally became SHA3, and so SHA3 hashing has changed.

We need to implement a new Keccak hash (AlgoName "KECCAK-256" I think, class name bouncycastle class name KeccakDigest)
and let EthHash refer to that rather than SHA3_256! We'll have to be careful whenever we use SHA/KECCAK hashes to be sure
we are using the right version. Grr. For now we're staying at 1.51 to keep our tests working.